Update module names
