+++ mcp-server/src/tools/mod.rs
@@ -1,4 +1,4 @@
 mod org_file_list;
 
 #[cfg(test)]
-mod tests;
+mod tools_tests;
