+++ org-core/src/lib.rs
@@ -2,7 +2,7 @@ pub mod error;
 pub mod org_mode;
 
 #[cfg(test)]
-mod tests;
+mod error_tests;
 
 pub use error::OrgModeError;
 pub use org_mode::OrgMode;
